// selfunder_template-final/selfunder_template-yosriraboudi/src/app/models/user.model.ts
export interface User {
  idUser: number;
  nom: string;
  prenom: string;
  email: string;
  password: string;
  telephone: string;
  photo?: string;
  dateNaissance: string;
  role: string;
  cin: number;
  adresse: string;
  profession: string;
  salaire: number;
  numTel: number;
  matriculeFiscale: string;

}